Temporary fix for the circular dependency
